separate files
Put import yang models functionalities into different files to reduce the size of extension.ts
